Treat unassigned site_diff_score as localized
This may be misleading by itself, but the true localization state (i.e. has been through localization and is localized, rather than just hasn't been through the localization process) can be determined by the combination of this boolean value and the site_diff_score field, in the event that this distinction is important.